1、完成websocket行情api的封装库 2、完成rest交易api的框架

+++ HBAPI/source/msg/msglastkline.cpp
@@ -35,7 +35,7 @@ bool MsgLastKLine::ReceiveJson(const QJsonObject& joPayload)
 
 	ptLastKLineData->eSymbolId = GetSymbolIdType(joPayload[szAttributeName[AN_SYMBOLID]].toString());
 	ptLastKLineData->ePeriod = GetPeriodType(joPayload[szAttributeName[AN_PERIOD]].toString());
-	ptLastKLineData->tTime = joPayload[szAttributeName[AN_TIME]].toInt();
+	ptLastKLineData->tTime = joPayload[szAttributeName[AN_TIME]].toDouble();
 	ptLastKLineData->dPriceOpen = joPayload[szAttributeName[AN_PRICEOPEN]].toDouble();
 	ptLastKLineData->dPriceHigh = joPayload[szAttributeName[AN_PRICEHIGH]].toDouble();
 	ptLastKLineData->dPriceLow = joPayload[szAttributeName[AN_PRICELOW]].toDouble();
